move config to config package
